Separate the functions as modules
